requirements.txt
python-telegram-bot
gspread
google-auth
python-dotenv
openai
